Conversion: zRefernceProcessor.cpp
Converting the soft_ref_delay() from uint64_t to jlong converts from
unsigned to signed, but as long as the value is less than 2^63 (very
large) the conversion is fine.

+++ src/hotspot/share/gc/z/zReferenceProcessor.cpp
@@ -47,7 +47,7 @@ class ZLRUMaxHeapPolicy : public LRUMaxHeapPolicy {
 public:
   // Capture state (of-the-VM) information needed to evaluate the policy
   void setup() {
-    _max_interval = ZAdaptiveHeap::soft_ref_delay();
+    _max_interval = (jlong)ZAdaptiveHeap::soft_ref_delay();
   }
 };
